Don't use pairs in Polygon:triangulate
Polygon:triangulate calls isEar on each iteration through the vertices, isEar calls anyPointInTriangle which does a pairs loop.

Fixed by renaming anyPointInTriangle and to noPointIntriangle, and returning the index of q when there are no points in the pqr triangle.

This allows us to use a list instead of a set!